refactor(clients)!: Rename OSV classes according to ORT conventions
OSV currently is the only client library project that provides another
"wrapper" around the low-level service API. Use "ServiceWrapper" as a
conventional suffix for that class.

Reword some docs / rename some variables accordingly along the way.
